refactor(engine): move interface to engine
Only the engine needs this interface at the moment
